+++ src/Spy/index.ts
@@ -1,0 +1,3 @@
+export {CallsCounter} from "./CallsCounter";
+export {SpyInterface} from "./SpyInterface";
+export {Visitor} from "./Visitor";